投稿とコメントに評価を行える処理追加

+++ src/app/api/user/delete/route.ts
@@ -8,13 +8,14 @@ import z from "zod";
 // ユーザー削除
 export async function POST(req: NextRequest) {
 
-    const cookies = await getCookies();
+    const cookie = await getCookies();
 
-    if (cookies === null) {
+    if (cookie === null || cookie === undefined) {
         return NextResponse.json({ message: "ログインしていません。", success: false, login: false }, { status: 401 });
     }
-    const userId = cookies.id;
-    const id = 0;
+
+    const userId = cookie.id;
+    const { id } = await req.json();
 
     try {
         // ログイン判定
@@ -33,13 +34,16 @@ export async function POST(req: NextRequest) {
             return NextResponse.json({ message: message, success: false }, { status: 400 });
         }
 
+        if (Number(userId) !== Number(id)) {
+            return NextResponse.json({ message: "権限がありません。", success: false }, { status: 403 });
+        };
+
         // ユーザー削除
         await prisma.users.delete({
-            where: { id: Number(userId) }
+            where: { id: Number(id) }
         });
 
         return NextResponse.json({ message: "ユーザーを削除しました。", success: true }, { status: 200 });
-
     } catch (e) {
         return NextResponse.json({ message: "ユーザー削除 失敗...", e }, { status: 500 })
     }
